minor bugfixes; removal of libboost_filesystem

+++ include/Typedef.h
@@ -2,6 +2,6 @@
 #define TYPEDEF_H
 
 // define floating point format for (almost) all data used in shaplens
-typedef float data_t;
+typedef double data_t;
 
 #endif
